Merged PR 761441: Improve fire-forget materializeoutput feature
In the "Fire & Forget MaterializeOutput" feature of BuildXL,  the scheduler attempts to send 'MaterializeOutput' requests to each unattached workers. Because those workers are not connected to the orchestrator yet, we wait for a task (m_beforeSendingToRemoteTask) in SendToRemote method before we add 'MaterializeOutput' requests to the list of builds requests (m_buildRequests). These requests are subsequently dispatched to workers through a thread named 'm_sendThread'. The 'm_beforeSendingToRemoteTask' waits for either worker attachment or a combination of the scheduler's completion (excluding MaterializeOutput steps) and a default timeout of one minute. If attachment is prolonged, the process typically waits for the scheduler's completion, as the timeout is relatively short. In a highly cached MetaBuild scenario, all steps except MaterializeOutput can be executed swiftly, often within a minute, making the timeout a critical factor. A prolonged attachment, exceeding a few minutes, hinders workers from connecting to the orchestrator and participating in the ProductBuild stage. While increasing the timeout is a simple solution, I have further optimized this process.

Now, the scheduler still attempts to send 'MaterializeOutput' requests to each unattached worker, but it no longer waits in the SendToRemote method. Instead, it directly adds these requests to 'm_buildRequests', eliminating numerous dangling tasks. In the 'm_sendThread' thread, prior to dispatching a batch of build requests to a worker, the system waits for either the worker's attachment or the combined event of scheduler completion and the timeout. This modification ensures that the MaterializeOutput step does not get blocked due to unattached workers. Once the MaterializeOutput step is completed for the local worker for each pip, the scheduler completes. However, pending build requests already exist in the unattached workers. Therefore, the task awaited in SendToRemote becomes the scheduler's own completion, which now includes the MaterializeOutput operations for the local worker. This revised approach, even with the same timeout duration, allows workers more time to connect to the orchestrator.

Related work items: #2138254