Merged PR 761892: Use ADO environment variables to infer cache miss analysis keys based on branches related to the build
Retrieve contextual information about the branch the build is running from the Azure DevOps environment variables, and use those values as keys. We consider two scenarios: builds running from a PR trigger and builds that don't.

For builds running on a PR, these are the candidate keys we use:
1. The PR merge-branch name (e.g., /refs/pull/1234/merge)
2. The source branch for the PR (e.g., /refs/dev/chelo/myFeatureBranch)
3. The target branch for the PR (e.g., /refs/heads/main)

The rationale is that we want to get the latest fingerprint store pushed by other iterations from 'this' PR, which will share the PR branch name, we fall back to the closest branches that might have been built before in a non-PR build (and thus pushed a fingerprint store using their branch names).

For builds triggered outside or PRs, we just use the branch name, which is BUILD_SOURCEBRANCH. The other variables related to PRs are undefined in this context.

This strategy is effective assuming the target branches are built regularly and that FP stores from those branches will become accessible from the PR builds (i.e., with 'baseline builds' running in branches that are targeted from PRs, and in the same cache namespace that the PRs).