Merged PR 761819: Handle CacheInitializationtask failure for dev builds
GetCacheInitializationTask method returns a task, result(failure if present) for a distributed build are handled instantly. But in case of dev builds we handle them during GraphReuse and ConstructingSchedule.
Depending on the stack trace, the issue occurs when the initialization of cache has failed, but the ConstructSchedule method has decided to proceed with new graph and the method returns successfully.
But an error is logged if there is any failure in the initialization of the cache. This causes the exception we see as it fails the assertions in ValidateSuccessMatches method.

Related work items: #2138736