Merged PR 763122: Include current fingerprint salt when calculating static fingerprint during graph construction from graph fragments
The bug is caused by the following scenario:

Consider a scenario where a build session consists of 2 builds, one consisting pips for constructing pip graph fragments, and the other stitches the fragments together and executes the pips in the fragments. In the first build session, the 1st build has a pip that constructs a pip P in a graph fragment G with a fingerprint salt A. The fragment is then included in the 2nd build that executes P.

In the next build session, the first build keeps using the fingerprint salt A for constructing the graph fragment G. Thus, in that build, the construction of G has a cache hit. In the second build, one changes the fingerprint salt to B, and expects P to have a cache miss. In this build we will have a graph cache miss because we have a different fingerprint salt B. However, during the graph construction, we still serialize the graph fragments generated in the first build. If we do not include the fingerprint salt B when calculating the pip's static weak fingerprint (when stitching the fragments together), then P will have the same static weak fingerprint as the one in the first build session. This can result in a hit in the incremental scheduling state and make P get skipped during the execution phase, and thus underbuild.

Related work items: #2137467