Merged PR 763628: [Detours] Fix path canonicalization for near-root paths
For file access reports, Detours in general normalize/canonicalize the paths by calling `GetFullPath`.

However, there is one case where Detours does not do that, i.e., the case where it removes path last segment.

Consider the following API call, `FindFirstFile("B:\\*.cpp")`. This is a call to start enumerating `B:\` with `*.cpp` pattern. Detours needs to report enumerate access on `B:\`. To do so, it has to remove `*.cpp` from the given input. The last segment removal was so naive that it simply removes everything after, and including, the last `\`. This results in Detours reporting `B:`.

The fix is to compare the path with the path root. If the position of last `\` is within the root's length, then we should take the whole root instead.

Related work items: #2126559