Merged PR 757744: [Ephemeral] Track Azure locations within the content tracker
The ephemeral cache currently assumes that something exists in Azure Storage if it exists in the content tracker (distributed, accessible to all machines) OR if it's in the put elision cache (local to the machine). This causes a series of things to happen:
1. The number of storage accesses is higher than it should be in cases where the content is added in one machine, evicted, and then accessed from another. This scenario is relatively common because the local cache size we use for ephemeral builds is small as we have small drives.
2. A race condition is possible (see deleted comments); it's prevented by waiting, but the cost of that is more storage accesses.
3. In the datacenter wide case, it also means that the accessing of Azure Storage is less efficient than it could be, because the knowledge about storage accesses isn't shared across builds.

This PR adds Azure Storage as a location and modifies things so they take that extra information into consideration.

Example builds: https://cloudbuild.microsoft.com/user/jubayard_20231215164957
Further example builds: https://cloudbuild.microsoft.com/user/jubayard_20240125180851

Related work items: #2122872