Merged PR 763374: Enable uploading cache logs to Kusto
Reuses the NLog component already available for CASaaS. The schema of the kusto table is preserved to make existing kusto queries compatible, even though many of the columns are not meaningful outside of CloudBuild.
For the time being, adding an explicit flag to turn it on for deployment purposes, but eventually the existing /logsToKusto flag will also turn on cache logging.

Instructions/scripts for setting up the ingestion/table are coming afterwards.