Merged PR 765084: Make violation log more generic to account for dynamic violations not reporting the proper path
For some observations we report the root process of the pip upon encountering an access violation, and this can be misleading to a user that wants to allow-list the access based on process name. This has happened in the past. This PR makes the message more generic so someone reading the message doesn't jump easily to conclusions about which specific process performed the access.