Merged PR 765572: Tidy up method for setting execute file permission
The use of Try* prefix for `TrySetExecutePermissionIfNeeded` is confusing. Typically, Try method returns simply `bool` where the `false` value indicate that the action is not successful.

In the case of `TrySetExecutePermissionIfNeeded`, the return value of `true` and `false` have some meanings, i.e., whether `chmod` may be called or not. So, the use of Try prefix is inappropriate here.

Also:
- Simplify method signatures in LocalDiskContentStore.