Merged PR 766475: Attempt to address BuildXL crashes due to running out of channel space
BuildXL is crashing because we're either producing too many logs that don't seem to be flushed down quickly enough, or the task that flushes them is dying. This PR increases the barrier to a crash, and logs if the task crashes for any reason.

The aim is to get some information about when this happens. We'll also be decreasing the amount of logging we do.