Merged PR 766489: Start remote workers only if we actually have pips to run
When the workers are initialized we start the `m_sendThread` inside `RemoteWorker`. If then the filter returns no nodes, we just skip all the scheduler logic and dispose the engine: this goes on to dispose the `RemoteWorker`s, and here `Dispose` asserts that send thread is not alive. So we shouldn't start them before we have a chance to check that we won't actually run a build