Merged PR 766983: Linux test: print debug/error message only when assertion failed
The stdout/stderr from the unit test may contain character that xunit doesn't like.

The proper fix should be in xunit, but according to https://github.com/xunit/xunit/issues/865, this should have been fixed. I'm not sure why the fix still doesn't work.

One can think of this as an attempt to avoid possible retries implemented in this PR !766918