Merged PR 767263: Demote PipOutputUpToDate to Diagnostic level
This is the event that is logged when an output on disk is reused due to already being in the correct state. This has a very large impact on logging volume in Kusto for CloudBuild builds.

I decided to remove it from local logs as well to keep consistency between what's logged to Kusto and what's in the file log. In CloudBuild, the Kusto log is probably more valueable than that local file log anyway. In desktop builds it is usually not going to be logged due to incremental scheduling. So it is already inconsistently logged.

We can always add this back into local file logging if we end up missing it. But this is a good opportunity to be a little more aggressive about shrinking our logs all up.