Merged PR 768178: UTs showing underbuild due to directory probe
Underbuild can happen when a existing directory probe is not considered as an input by ObservedInputProcessor.

This can happen if
1. the probe itself is not reported by Detours, or
2. the probe itself is not considered as probing source directory.

For (1), directory probe is only reported if (a) the flag for reporting directory probe itself is enabled, and (b) the probed directory is under an input directory (i.e., a directory dependency).

For (2) , even if the directory probe itself is reported, but if ObservedInputProcessor says that the directory is not a source directory, then it is assumed as a non-existing directory. This can happen if the probed directory is not in the scope of the directory dependency, e.g., the probed directory is A/B/C, but the source sealed directory is only top-level of A.

Related work items: #2146397