Merged PR 767614: Modified the build status line for IPC pips progress
Modified the ConsoleEventListener.cs to completely remove the service pips counter from the build status line.
Displays the message - Waiting for background tasks to complete
when the fancyConsole option is both disabled/enabled or for ADO builds if these conditions are satisfied
a.) The build is complete.
b.) When the service pips are still running.

![fdao.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/767614/attachments/fdao.png)

![nonfancyyyy.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/767614/attachments/nonfancyyyy.png)

![fann.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/767614/attachments/fann.png)

Related work items: #2138416