Merged PR 770360: Add a lage location to the lage resolver
This allows users to specify the lage location directly for the corresponding resolver (in addition to the current `npm run lage` approach). This opens the door for customers to specify particular versions of Lage for bxl to use without disrupting the 'lage' script defined in package.json, which may be used by developers outside of bxl.

Related work items: #2154615