Merged PR 771473: Additional handling for internal error cancellation crashes
The code leading up to TryCreateTwoPhaseCacheEntryAndStoreMetadataAsync() stores each output file to the cache. Prior to that there is a short circuit for cancellation, but I suspect that storing all of those outputs can be a long operation for large pips. Add another check prior to storing metadata to ensure that is not attempted if the scheduler has been cancelled.

Related work items: #2156570