Merged PR 771256: Fail MetaBuild on the worker if some materialize output requests fail to be sent
I introduced this bug when redesigning the attachment logic. When some materializeOutput requests cannot be sent due to various reasons and the connection is still healthy, the worker still proceeds to the product build. It causes 'files not found' errors during product build on that worker.

It roughly affects 5-10 builds daily.

Related work items: #2148861