Merged PR 771829: Relax config calls so arbitrary expressions are accepted
We were enforcing that the argument that gets provided to config( ... )  in the main config file has to be a literal. This blocks cases like

`config(importFile(f`\`satelliteconfig.dsc\`).buildMyLiteral( ...some other params... ));`

The restriction is pretty arbitrary, and offloading the argument to a different function is useful for factoring out common logic across different config files (e.g. linux vs windows).

This PR relaxes that restriction so we allow any expression. The type checker still makes sure that types match.