Merged PR 771777: Add read-only mode to the blob cache client.
Expose an 'IsReadOnly' flag to the blob cache factory. This can later be used via the vertical aggregator to configure a dev cache where the remote is only used for pulling.
The 'last pinned time' optimization performance for service-less GC is turned off on this case. This means pulling a CHL does produce preventive pinning, but the last pin time is not updated. Subsequent pulls will preventively pin again. This is unlikely to move the perf needle, as long as publishing builds run within the given retention period (default 6 days), the preventive pinning won't happen (if publishing builds run less frequently than that, the chances of cache hits are pretty slim anyway)
Read-only mode is not supported yet for the ephemeral cache. It could be added in the future, but it is unclear than an ephemeral cache will be need for dev cache (dev side).

Related work items: #2145607