Merged PR 765776: Add internalWarnings to DominoCompletion event
Warning events containing the 'InfrastructureError' keyword will be recorded throughout the build. A maximum of five unique warning events will be saved. At the end of the build, these warning events will be logged as a part of the DominoCompletion event. The objective is to identify internal or infrastructure-related issues within green builds. Such builds are considered sub-optimal, and it is important to monitor the frequency of these issues and receive alerts when necessary.