Merged PR 772031: Update APIScan yml to avoid using secret
Retrieve workload federation token in azure client task, and use this token to log in az.
AzureServiceTokenProvider uses AzureCli to get token. The account it use is the service principle logged in in previous step.
This way we avoid using secret to do authentication. APIScan use azure client and the service principle as the user to do authentication. The azure login step is using workload federation.

Related work items: #2136773