Merged PR 773240: Fix logging and exit conditions when Hello to orchestrator fails
Currently, the attach timeout is also used deep into the gRPC call, in the "wait for connection" section. So when the orchestrator is absent from the start, this `Wait()` of the Hello call is not really timing out, rather completing with a failed state. This makes us return `true` in `SayHello`, so we continue with "Wait for attachment", but the build is now exiting. So we end up logging this:

```log
[0:03.248] [19:39:46.67 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] Grpc.NET auth is enabled.
[0:03.267] [19:39:46.69 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] Connecting by Hello
[0:03.286] [19:39:46.71 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] Idle -> Connecting
[0:03.302] [19:39:46.73 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] Connecting -> TransientFailure
....
[20:03.271] [19:59:46.69 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] TransientFailure. Failed to connect in 1200004ms. Failure A task was canceled.
[20:03.274] [19:59:46.70 UTC] verbose DX7029: [88429e44c000000.internal.cloudapp.net] TransientFailure -> Shutdown
[20:03.284] Waiting for orchestrator to attach
[20:03.285] [19:59:46.71 UTC] error DX7004: There were no calls from orchestrator in the last 20 minutes. Assuming it is dead and exiting.
```

The last two lines in particular are wrong to log and only happen because of what I just described - if SayHello returned false, we would skip this logic and just exit.

This PR bubbles up the failure in Hello so we log an appropriate message and never "wait for attachment":

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/773240/attachments/image.png)

Note that it's still appropriate to `Wait()` the hello call, because it could time out _after_ the "wait for connection"
 has succeeded.