Merged PR 773341: Add codespaces authentication support to the blob cache client
Codespaces uses the extension https://github.com/microsoft/ado-codespaces-auth to talk to the Microsoft AAD via Visual Studio Code and provide AAD authentication. This PR adds the ability for the blob cache client to interact with this extension and provide authentication when available. This is added as yet another auth mechanism the cache client supports.

Related work items: #2157484