Merged PR 773253: Add option to copy input paths under opaque directory for qtest pips
Currently we can only add file nodes or directory nodes in QTestRunner to copy into qtest sandbox. If we were to copy files under opaque directory, there does not seem to be an option in `runQTest` method to do it. This PR adds an option to copy input paths under opaque directory for qtest pips. The caller would also need to add the opaque directory node under qTestrunTimeDependencies as well to prevent DFA.