Merged PR 773589: Ensure directory symlink is tracked as a read when there is read through it
The change in this PR makes sure that directory symlink is tracked as a read when there is a read operation through it.

The added UT demonstrates the issue when there are both read and probe operations; the latter makes the directory symlink being treated as a directory, and thus is not tracked as a read.

To fix the bug and to achieve consistency between Detours and the managed sandbox code, the PR makes sure that the reporting in Detours uses a correct operation context. This allows us to have the same logic, in Detours and in sandbox, for determining if an operation is applied on a file or on a directory.

Other changes in this PR are cosmetics.

Related work items: #2154637