Merged PR 774386: FIx for the ADO console logs on the workers.
The counters used for the orchestrator do not represent the accurate values on the worker machines. They are not calculated as expected.
In order to avoid the confusion, modified the console status to only display the process pips executing on the worker.
This is exclusive to the workers on the ADO.

The ADO logs on the orchestrator machine remain unchanged
https://dev.azure.com/mseng/Domino/_build/results?buildId=27389793&view=logs&j=e7020543-e8ed-5e17-1ad4-e7bb89160f58&t=a127a243-c15c-5e7c-8982-d28f3304a152&l=81

The worker logs now only display the process pips executing
https://dev.azure.com/mseng/Domino/_build/results?buildId=27389793&view=logs&j=48ccc790-4aca-541d-f3d1-13fafd4ad28f&t=47bd7b74-57cd-522f-e5dc-4f4c398d9f57&l=73

Related work items: #2153520