Merged PR 774392: Remove the warning message generated when GenerateBuildManifest flag is not set with the expected value.
Added a flag called GenerateBuildManifest flag on the CB side.
This flag is enabled by default for all builds roles except buddy builds. But is enabled for windows buddy builds as well.

Since we have a couple of scenarios where we do not expect this flag to be present. Displaying this message for the buddy builds for example might be confusing to the customers.
Hence removing it.

Related work items: #2156237