Merged PR 774539: Delete some build outputs to speed up SDL binary analysis
This deletes ~40% of the outputs that SDL binary analysis needs to act upon. A quick test shows that it shaves about 3 minutes off of antimalware. It should to a lesser extent help on the other SDL tasks.

I expect it can be expanded for more improvement. I want to attack that from the producer side before I go too deep here.