Merged PR 774587: [Detours] Invalidate path cache for file symlink
This PR fixes the following bug:

1. File symlink is about to be created by first opening a handle `CreateFile(..., WRITE, ...)`.
2. File symlink is created by passing the handle from (1) to `DeviceIoControl(..., FSCTL_SET_REPARSE_POINT, ...)`
3. The target of file symlink is read though the symlink itself.

There's no access reported for the target because in step (1) the path of the file symlink is inserted to the cache as a non-symlink (yes, because it has not been created). Thus in step (3), the chain of reparse points are not enforced.

The change is simple: also invalidate path cache for file symlink (not only for directory)

Related work items: #2156501