Merged PR 776791: Use materialization info to check for reparse point in Observed input processor
Previously, observed input processor (OIP) directly probes the real file system to check if a path is a directory symlink or not.

The directory symlink is produced by an upstream pip. However, when minimal materialization is enabled, the directory symlink may not have been materialized when the consuming pip is running OIP for cache lookup.

For example, a pip P produces a directory symlink s, and another pip Q probes it. In the first run, Q will have an existing directory probe.

Before the second run, s is deleted.

In the second run, P has a cache hit, but s is not materialized due to being lazy. Now, when Q runs OIP for cache lookup, it doesn't see s in the real file system, and thus it categorizes it as absent path. Q then has a cache miss and executes.

In this PR, besides checking the real file system, OIP will include the materialization info registered by the upstream pips in the file content manager.

In the above example, when P declares that it has a cache hit, it registers all materialization info of its outputs to the file content manager. This materialization info include information about reparse point data.

Related work items: #2165434