Merged PR 776770: Fix log level for 'will retry pip on the orchestrator' message
If we fail to send a pip request to the worker three times, we should retry on the orchestrator. However, what was intended to be a warning event was mistakenly logged as an error. This was a bug. I have corrected this, changing the error to a warning. Now, if a pip fails to be sent to the worker three times, we should be able to retry on the orchestrator. This issue may be related to distribution challenges, such as sending millions of hashes to the worker.