Merged PR 777438: Make sure reports to the FIFO path get ignored
When ptrace is started there is a small overlap with the still existing interpose sandbox that makes ptrace observe some of what the interpose sandbox is doing. This is usually harmless, but this could involve interacting with the FIFO path when sending reports. This results in the FIFO path potentially becoming part of the pip fingerprint, and when bxl tries to hash it, we crash (since the FIFO file is not a real file).

Make sure we ignore these reports. The race is hard to avoid, and turning off interposing is not easily achievable either. Turning off reporting also has some undesirable side effects, like disabling debug logging on that part of the codebase (when this happens, is hard to understand why).

This PR ignores this report on managed side, so we don't slow down the process doing path comparisons. We also move some other path comparison logic (previously done on native side), to managed side as well.

Related work items: #2165745