Merged PR 777416: Avoid auto flush std input when setting up pipes for detoured processes
When setting up a StreamWriter for stdin in detoured process, the writer is set to auto flush. StreamWriter immediately does a write operation to the pipe (although the data is empty). However, when the pipe is closed (or is being closed), the write operation would fail and throw an exception. This can happen if the detoured process does not need any standard input, and the process terminates quickly before this pipe setup is completed.

It was not an issue before .NET 8 because when writing to a FileStream that represented a closed or disconnected pipe, the underlying operating system
error was ignored and the write was reported as successful. However, nothing was written to the pipe. Starting in .NET 8, when writing to a FileStream whose
underlying pipe is closed or disconnected, the write fails and an IOException is thrown.

See breaking changes in .NET 8: https://learn.microsoft.com/en-us/dotnet/core/compatibility/core-libraries/8.0/filestream-disposed-pipe

Related work items: #2166469