Merged PR 777640: Ensure pip validity check requiring graph happens in graph builder
Some pip validity check that requires graph happens in process builder. The problem is process builder is not required to have a graph. So, for such a check, the check should happen in pip graph builder instead; and this PR is about that.

And now, "Trusting statically declared accesses is not compatible with declaring opaque or source sealed directories" will have a proper error message.

Related work items: #2166157