Merged PR 773297: DumpPip Analyzer /includeStaticMembers flag
Update the DumpPip Analyzer to have the `/includeStaticMembers` option to enumerate statically known sealed directory members.  This is usefully when trying to figure out what files a process PIP depends on from other PIPs that produce those files. In addition the kind for each directory dependency is now included in the output.