Merged PR 778634: Add a marker file to incomplete drops
Artifacts' infra only allows two states for drops - finalized and non-finalized. To avoid having too many unfinished drops, we make sure that finalize command is run for all drops made by BuildXL. So, even if a build failed and some content was missing from the drop, we would still finalize it. This means that a consumer cannot tell if a drop is complete or not by just looking at the drop content and/or its metadata. Some builds (like Cosine) depend on drops made by other builds, i.e., they need to know if a specific drop is good to use. Before this PR, users solved this problem by forcing build manifest creation and checking its status (the manifest is only created for complete drops). Manifest creation is somewhat flaky (because of signing service being flaky), so this workaround resulted in them having fewer successful builds.

This PR makes DropD add a special marker file to incomplete drops that consumers can check. If such a file cannot be added, as a backup signal, DropD will leave the drop in a non-finalized state.

Related work items: #2167925