Merged PR 780205: Fix hanging when the scheduler is early-terminated with the internal error
When the scheduler is early-terminated due to an internal error, the workers that are released early can hinder the scheduler's shutdown, potentially causing the build to stall. Although the orchestrator receives results from these workers, the tasks assigned to the early-released workers cannot be completed if the scheduler on the orchestrator has already finished. Therefore, it is necessary to manually finalize the 'DrainCompletion' task sources for each worker.

Related work items: #2169589