Merged PR 782403: Refine logging ReceivedReportFromUnknownPid to avoid synthetic accesses
All occurrences of ReceivedReportFromUnknownPid from 1JS in the last week was about synthetic accesses (process requires ptrace, first allow write check in process). Let's not log in that case, we send these from native side before the process start event, and they are not real accesses in the end.

I'll check again in a week if that makes ReceivedReportFromUnknownPid reach 0, and in that case I'll turn it into a warning.