Merged PR 780371: Allow succeedfast pips to be recursively dirtied
There is logic added to prevent downstreams of succeedfast pips to be recursively dirtied. This can potentially cause underbuilds with incremental scheduling particularly when the upstreams of the pips in buildcone connected via succeedfast pip were dirtied.

The intent for the previous change is perf driven, to reduce the cache checks. However due to Eric's change which stops cache checks at the first up-to-date pip, the perf is retained and the change is not needed anymore. This PR removes the logic to stop downstreams of succeedfast pips to be dirtied.

Related work items: #2170157