Merged PR 783565: Revert "Merged PR 778027: Catch consolenotconnected exception and exit buildxl
The break is the change in StandardConsole.cs. That condition needs to be run any time the previous line was overwriteable to blank out the previously written text and move the cursor back to the appropriate position. The change was conditionalizing it when both the previous line was overwriteable and when the current line is.

I'm reverting the entire commit instead of just that one part because the error handling changes in the other files are related.

Related work items: #2138734, #2174322