Merged PR 784276: Fall back on simple console output on IOException
There are a handful of isolated crashes on linux and codespaces where an IOException is thrown on Console.CursorLeft. It's unclear what's causing this. This change takes the approach of encapsulating all of the complicated console behavior for updating the same lines in place. If an exception is thrown anywhere inside of that, it will stop attempting that more complicated functionality and fall back on a simple Console.WriteLine for all subsequent messages. This should give a much better chance of the build being able to continue when something odd happens.

Related work items: #2138734