Merged PR 784655: Revert "Merged PR 784360: Revert "Merged PR 784187: Remove net7 qualifier
Reapply the original Net7 removal commit `4c5cc753`. The issue that we initially saw after it was merged was not caused by the commit. The commit forced a cache miss for a couple of pip and it uncovered the issue with vctip/lib.exe.

Must be merged after !784653

Related work items: #2174274